requirements.txt
# Core LLM & API
groq>=0.9.0

# Web framework (FastAPI UI)
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
jinja2>=3.1.4

# Validation & settings
pydantic>=2.7.0
python-dotenv>=1.0.1
